ENT-4816 Do not create CFE_ROBOT user on upgrade

+++ packaging/common/cfengine-hub/postinstall.sh
@@ -914,12 +914,14 @@ $PREFIX/httpd/bin/apachectl start
 #Mission portal
 #
 
-true "Adding CFE_ROBOT key"
-( set +x
-  CFE_ROBOT_PWD=$(dd if=/dev/urandom bs=1024 count=1 2>/dev/null | tr -dc 'a-zA-Z0-9' | fold -w 32 | head -n 1)
-  $PREFIX/httpd/php/bin/php $PREFIX/httpd/htdocs/index.php cli_tasks create_cfe_robot_user $CFE_ROBOT_PWD
-)
-true "Done adding key"
+if ! is_upgrade; then
+  true "Adding CFE_ROBOT user"
+  ( set +x
+    CFE_ROBOT_PWD=$(dd if=/dev/urandom bs=1024 count=1 2>/dev/null | tr -dc 'a-zA-Z0-9' | fold -w 32 | head -n 1)
+    $PREFIX/httpd/php/bin/php $PREFIX/httpd/htdocs/index.php cli_tasks create_cfe_robot_user $CFE_ROBOT_PWD
+  )
+  true "Done adding user"
+fi
 su $MP_APACHE_USER -c "$PREFIX/httpd/php/bin/php $PREFIX/httpd/htdocs/index.php cli_tasks migrate_ldap_settings https://localhost/ldap"
 
 $PREFIX/httpd/php/bin/php $PREFIX/httpd/htdocs/index.php cli_tasks inventory_refresh
